requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.1
astroid==2.4.2
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.0.7
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.11.2
django-debug-toolbar==1.10.1
django-filter==2.4.0
django-import-export==2.5.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.8
isort==5.4.2
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
oauthlib==3.0.1
odfpy==1.4.1
openpyxl==3.0.7
pep8==1.7.1
Pillow==7.1.2
protobuf==3.15.3
pycodestyle==2.5.0
pylint==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.4.1
reportlab==3.5.56
requests==2.21.0
requests-oauthlib==1.2.0
six==1.15.0
sqlparse==0.3.0
stripe==2.27.0
style==1.1.0
tablib==3.0.0
toml==0.10.1
update==0.0.1
urllib3==1.24.2
virtualenv==20.4.2
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
